docker_requirements.txt
gpxpy
pathlib
pandas
numpy
pyyaml
bs4
matplotlib